BOCM-349 - fixes bug with merging arrays of appeals
